Refactor OBD admin helper boundaries
