+++ apps/server/vibesensor/adapters/obd/admin_runtime.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from vibesensor.adapters.obd.admin_client import ObdAdminClient
 from vibesensor.adapters.obd.admin_state import observe_configured_obd_device
-from vibesensor.adapters.obd.runtime_admin_state import ObdRuntimeAdminState
+from vibesensor.adapters.obd.runtime_store import ObdRuntimeStore
 
 __all__ = ["ObdAdminRuntime"]
 
@@ -12,21 +12,27 @@ __all__ = ["ObdAdminRuntime"]
 class ObdAdminRuntime:
     """Refresh configured-device admin state without exposing raw client actions."""
 
-    __slots__ = ("_admin_client", "_runtime")
+    __slots__ = ("_admin_client", "_store")
 
     def __init__(
         self,
         *,
         admin_client: ObdAdminClient,
-        admin_state: ObdRuntimeAdminState,
+        store: ObdRuntimeStore,
     ) -> None:
         self._admin_client = admin_client
-        self._runtime = admin_state
+        self._store = store
 
     def refresh_configured_device(self) -> None:
-        configured_mac = self._runtime.configured_device_mac_snapshot()
+        with self._store._lock:
+            configured_mac = self._store.policy.configured_device_mac
         observation = observe_configured_obd_device(
             admin_client=self._admin_client,
             configured_mac=configured_mac,
         )
-        self._runtime.apply_admin_observation(configured_mac, observation)
+        with self._store._lock:
+            self._store.state.apply_admin_observation(
+                observed_configured_mac=configured_mac,
+                current_configured_mac=self._store.policy.configured_device_mac,
+                observation=observation,
+            )
